https://trello.com/c/n1C8YVK4/132-it-is-not-a-bug-but-i-do-not-like-the-way-animation-work-in-this-case

+++ WrapLive/Views/WLProgressBar/WLProgressBar.h
@@ -30,7 +30,7 @@
 
 @property (nonatomic, weak) WLUploadingItem *uploadingItem;
 
-@property (nonatomic) CGFloat progress;
+@property (nonatomic) float progress;
 
 @property (strong, nonatomic, readonly) UIView *backgroundView;
 @property (strong, nonatomic, readonly) UIView *progressView;
